requirements.txt
customtkinter
firebase-admin
pillow
opencv-python
python-dateutil
pandas
openpyxl
tkcalendar
pyfingerprint
pyserial
pywin32
